DELIA-49777: Added condition not to clear mode settings.
Reason for change: There were some race conditions where we are delting
                   flag even if the global ip is available.

Test Procedure: Build and Verify.

Risks: Low

(cherry picked from commit 34b15138eaea5d27128ed2dec2b43bc020f0de63)

+++ ipv6addressChange.sh
@@ -60,14 +60,20 @@ if [ "x$cmd" == "xdelete" ] && [ "x$flags" == "xglobal" ]; then
      if [[ $addr == fd* ]] || [[ $addr == fc* ]]; then
        echo " It is ULA address, no need to clear IPv6 flags"
      else
-       echo " Box was previously in IPv6 mode and now in IPv4 mode. Clearing IPv6 flags for $ifc"
-       rm -f  /tmp/estb_ipv6
-       rm -f /tmp/addressaquired_ipv6
+       globalip=`ip -6 a | grep global`
+       if [ "x$globalip" = "x" ]; then
+          echo " Box was previously in IPv6 mode and now in IPv4 mode. Clearing IPv6 flags for $ifc"
+          rm -f  /tmp/estb_ipv6
+          rm -f /tmp/addressaquired_ipv6
+       fi
      fi
    else
-     echo " Box was previously in IPv4 mode and now in IPv6 mode. Clearing IPv4 flags for $ifc"
-     rm -f /tmp/estb_ipv4
-     rm -f /tmp/addressaquired_ipv4
+       globalip=`ip -4 a | grep global`
+       if [ "x$globalip" = "x" ]; then
+          echo " Box was previously in IPv4 mode and now in IPv6 mode. Clearing IPv4 flags for $ifc"
+          rm -f /tmp/estb_ipv4
+          rm -f /tmp/addressaquired_ipv4
+       fi
   fi
 fi
